0.2.7 (#143)
* Delete .DS_Store

* Added role updates to log (#92)

* Maintenance Command (#93)

* Added Github Bug system

* Finaly closed #84

* Update Dev to newest (#98)

* Delete guilddata.txt

* Update .gitignore

* Update ISSUE_TEMPLATE.md

* Reformat GitBug

* Fully Replaced

* Added New Bug System (Fix #84) (#99)

* Connected bug command with Github issues

* SelfMention - Call delete

Deletes the @Rubicon

* Fixed Verification System and BotLeaveListener Bug (#104)

* Fixed #102

* Serverinfo

* New Bug System Improvments

* Fixed CommandMute Error (#109)

* Fixed #100

* Fixed Serverinfo and added Multiserver SUpport

Close #95

* Optimized Imports and Reformat Code

* Feedback Command and Bug Fixed

* Feedback Command and Bug Fixed (#119)

* Fixed #115

* Fixed #114

* Urban Command own Thread

Close #121

* Fixed Imports and Reformat Code

* Finally Fixed Verification

* Joke command

* Fortnite Command Created

* Stuff

* Ow Stats added

* f

* Gitbug

* Some Updates

- Updated RequestHeader.java
- Reworked whitelist and blacklist system (not finally in this commit)
- Updated VerificationListener
- Implemented bots.discord.pw
- Improoved SafeMessage.java

* Fixed #122

* Finished blacklist system

- Updated deprecated stugg
- Updated white and blacklist

* Some updates (#135)

* Some Updates

- Updated RequestHeader.java
- Reworked whitelist and blacklist system (not finally in this commit)
- Updated VerificationListener
- Implemented bots.discord.pw
- Improoved SafeMessage.java

* Finished blacklist system

- Updated deprecated stugg
- Updated white and blacklist

* Changed usage

- Changed wrong usage

* Added Mentions in say (#134)

* Delete .DS_Store

* Fortnite Command Created

* Stuff

* Fixed #122

* Ow stats back

* Update CommandBlacklist.java

* Ow stats complete

* Removed Deprecated Permission Stuff

* Threading

* fixed permissions

* permissions

* Flashbot Stuff

* Update

* Update (#137)

* IGNORE THIS

* Update

- Fixed some Deprecated stuff
- Improofed HttpRequest
- Added Minecraft Command for our team member @leedjd who loves Minecraft

* CONFLICTS!

- Fixed Conflixts

* Updated API

- Updated discordbots.org API

* Fixed #140

* 0.2.6

* Fixed Join- & Leavmessage Command

+++ src/main/java/fun/rubicon/features/VerificationKickHandler.java
@@ -77,8 +77,14 @@ public class VerificationKickHandler {
             if (!verifyKicks.containsValue(this)) return;
             Guild guild = RubiconBot.getJDA().getGuildById(this.guildid);
             Member member = guild.getMemberById(this.userid);
-            member.getUser().openPrivateChannel().queue(c -> c.sendMessage(this.kickText.replace("%invite%", guild.getTextChannelById(RubiconBot.getMySQL().getVerificationValue(guild, "channelid")).createInvite().setMaxUses(1).complete().getURL())).queue());
-            guild.getController().kick(member).reason(this.kickText).queue();
+            if (member.getUser().isBot()) {
+                verifyKicks.remove(this);
+                return;
+            }
+            if (guild.getSelfMember().canInteract(member)) {
+                member.getUser().openPrivateChannel().queue(c -> c.sendMessage(this.kickText.replace("%invite%", guild.getTextChannelById(RubiconBot.getMySQL().getVerificationValue(guild, "channelid")).createInvite().setMaxUses(1).complete().getURL())).queue());
+                guild.getController().kick(member).reason(this.kickText).queue();
+            }
             VerificationKickSQL sql = new VerificationKickSQL(member.getUser(), member.getGuild());
             RubiconBot.getJDA().getGuildById(guildid).getTextChannelById(RubiconBot.getMySQL().getVerificationValue(guild, "channelid")).getMessageById(Long.parseLong(sql.get("message"))).complete().delete().queue();
             remove();
